refactor: move parser.py to query.py
