+++ tests/data/unit_tests/test_query/test_case_statement_parsing.sql
