+++ tests/data/unit_tests/test_query/test_cte_parsing.sql
